pull out emitters into their own module
